+++ tap_mongodb/__init__.py
@@ -1,2 +1,2 @@
 """MongoDB tap."""
-__version__ = "2.2.0"
+__version__ = "2.3.0"